requirements.txt
aiogram==3.6.0
SQLAlchemy==2.0.36
asyncpg==0.29.0
python-dotenv==1.0.1
uvicorn==0.32.0
